Tidy up structure
