+++ services/orchest-webserver/client/src/lib/orchest/context.jsx
@@ -1,0 +1,9 @@
+// @ts-check
+import React from "react";
+
+/**
+ * @type {React.Context<import("@/types").IOrchestContext>}
+ */
+export const OrchestContext = React.createContext(null);
+
+export const useOrchest = () => React.useContext(OrchestContext);
